Add files via upload
Forecasting code explained in a markdown document. Accesses raw transaction data in JSON from the Clover API and uses this to create a linear regression model. The model is then used to create a rshiny app calculator which is being used by a restaurant to forecast sales on a daily basis.